[SDK] Default allow_dtype_casting to off
Disallow automatic dtype casting by default.
This feature makes the stack less predictable by introducing implicit
behaviour, and has already introduced several bugs including breaking
FP8 logits verification.
So just keep the feature turned off by default.

MAX_INTEGRATION_TESTS_REV_ID: 0180f0b1a75c7da6a3ebffcf65817580af4698ee

+++ max/tests/integration/API/python/test_speculative_decoding.py
@@ -172,7 +172,7 @@ def test_config__validate_huggingface_engine(smollm2_135m_local_path) -> None:
     ):
         PipelineConfig(
             model_path=smollm2_135m_local_path,
-            quantization_encoding=SupportedEncoding.float32,
+            quantization_encoding=SupportedEncoding.bfloat16,
             device_specs=[DeviceSpec.accelerator()],
             draft_model_path=smollm2_135m_local_path,
             engine=PipelineEngine.HUGGINGFACE,
